Change unsigned long to unsigned int in md5 code, so that it compiles and works on 64-bit architectures where unsigned long is 64 bits.

+++ fitz/fitz_stream.h
@@ -284,8 +284,8 @@ typedef struct fz_md5_s fz_md5;
 
 struct fz_md5_s
 {
-	unsigned long state[4];
-	unsigned long count[2];
+	unsigned int state[4];
+	unsigned int count[2];
 	unsigned char buffer[64];
 };
